Styled the page to look like dark mode which I prefer
Had so much trouble with styling the cards, ended up using chatgpt to help me with the css to center the list inside the box and make them stack horizontally.

Wasn't supposed to use flexbox because the css exercise hasn't reached that far yet but tbh i'm not sure if its doable without it.